Settings to suspense
